Move Python proto package into subdirectory
