+++ sdk/python/feast/infra/local_sqlite.py
@@ -6,9 +6,9 @@ from typing import Dict, List, Optional, Tuple, Union
 from feast import FeatureTable, FeatureView
 from feast.infra.key_encoding_utils import serialize_entity_key
 from feast.infra.provider import Provider
+from feast.protos.feast.types.EntityKey_pb2 import EntityKey as EntityKeyProto
+from feast.protos.feast.types.Value_pb2 import Value as ValueProto
 from feast.repo_config import LocalOnlineStoreConfig
-from feast.types.EntityKey_pb2 import EntityKey as EntityKeyProto
-from feast.types.Value_pb2 import Value as ValueProto
 
 
 def _table_id(project: str, table: Union[FeatureTable, FeatureView]) -> str:
