+++ sdk/python/feast/infra/provider.py
@@ -3,9 +3,9 @@ from datetime import datetime
 from typing import Dict, List, Optional, Tuple, Union
 
 from feast import FeatureTable, FeatureView
+from feast.protos.feast.types.EntityKey_pb2 import EntityKey as EntityKeyProto
+from feast.protos.feast.types.Value_pb2 import Value as ValueProto
 from feast.repo_config import RepoConfig
-from feast.types.EntityKey_pb2 import EntityKey as EntityKeyProto
-from feast.types.Value_pb2 import Value as ValueProto
 
 
 class Provider(abc.ABC):
