+++ sdk/python/feast/registry.py
@@ -22,10 +22,10 @@ from urllib.parse import urlparse
 
 from google.auth.exceptions import DefaultCredentialsError
 
-from feast.core.Registry_pb2 import Registry as RegistryProto
 from feast.entity import Entity
 from feast.feature_table import FeatureTable
 from feast.feature_view import FeatureView
+from feast.protos.feast.core.Registry_pb2 import Registry as RegistryProto
 
 REGISTRY_SCHEMA_VERSION = "1"
