+++ sdk/python/setup.cfg
@@ -4,7 +4,7 @@ include_trailing_comma=True
 force_grid_wrap=0
 use_parentheses=True
 line_length=88
-skip=feast/types,feast/core,feast/serving,feast/storage,feast/third_party
+skip=feast/protos
 known_first_party=feast,feast_serving_server,feast_core_server
 default_section=THIRDPARTY
 
@@ -13,7 +13,7 @@ ignore = E203, E266, E501, W503
 max-line-length = 88
 max-complexity = 20
 select = B,C,E,F,W,T4
-exclude = .git,__pycache__,docs/conf.py,dist,feast/core,feast/serving,feast/types,feast/storage,feast/third_party
+exclude = .git,__pycache__,docs/conf.py,dist,feast/protos
 
 [mypy]
 files=feast,test
