+++ sdk/python/tests/cli/online_read_write_test.py
@@ -2,8 +2,8 @@ from datetime import datetime, timedelta
 from pathlib import Path
 
 from feast.feature_store import FeatureStore
-from feast.types.EntityKey_pb2 import EntityKey as EntityKeyProto
-from feast.types.Value_pb2 import Value as ValueProto
+from feast.protos.feast.types.EntityKey_pb2 import EntityKey as EntityKeyProto
+from feast.protos.feast.types.Value_pb2 import Value as ValueProto
 
 
 def basic_rw_test(repo_path: Path, project_name: str) -> None:
