+++ sdk/python/tests/test_feature_table.py
@@ -20,11 +20,11 @@ import grpc
 import pytest
 
 from feast.client import Client
-from feast.core import CoreService_pb2_grpc as Core
 from feast.data_format import ParquetFormat, ProtoFormat
 from feast.data_source import FileSource, KafkaSource
 from feast.feature import Feature
 from feast.feature_table import FeatureTable
+from feast.protos.feast.core import CoreService_pb2_grpc as Core
 from feast.value_type import ValueType
 from feast_core_server import CoreServicer
